# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Example API",
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Root endpoint",
        "description": "Returns a welcome message",
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/access": {
      "get": {
        "summary": "Access data retrieval",
        "description": "Retrieves access data for a user",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Access data returned",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Access data submission",
        "description": "Submits access data",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Access data submitted",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Access data update",
        "description": "Updates access data",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Access data updated",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Access data deletion",
        "description": "Deletes access data",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Access data deleted",
            "schema": {
              "type": "object",
              "properties": {
                "Message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}